\name{grid-internal}
\alias{wnmf}
\alias{wpnmf}
\alias{entropy2}
\alias{recoverU}
\alias{recoverV}
\alias{somplot}
\alias{transit.prob}
\alias{rwr.mcmg}
\alias{metacell.expression}
\alias{ldpois}
\alias{add.grid}
\title{Internal dpath Functions}
\description{
 Internal dpath functions
}
\details{
These are not to be called by the user 
}
\keyword{ internal }

