\name{etv2}
\docType{data}
\alias{etv2}
\title{An example single cell RNA-seq data matrix}
\description{
This data set contains single cell RNA-seq expression data of 5,799 genes in 281 
cells.  This is a comprehensive single cell RNA-seq dataset of
Etv2-EYFP positive cells captured at E7.25, E7.75 and E8.25 during mouse development.
}
\usage{etv2}
\format{A numeric matrix containing 5,799 rows (genes) and 281 columns (cells)}
\examples{
library(dpath)
data(etv2)
}
\keyword{datasets}
