\name{reorder.dpath}
\alias{reorder.dpath}
\title{
Reorder the metagenes from the repetitive runs.
}
\description{
Reorder the metagenes from the repetitive runs.
}
\usage{
\method{reorder}{dpath}(x, ...)
}
\arguments{
  \item{x}{A \code{dpath} object}
  \item{...}{other arguments}
}
\value{
	A \code{dpath} object
}
\author{
Wuming Gong <gongx030@umn.edu>
}
\examples{
	help('dpath')
}
\keyword{dpath}
